requirements.txt
Django==4.2.9
django-environs==0.1.1
python-telegram-bot==13.15
environs==10.3.0
